requirements_tests.txt
-r requirements.txt
coverage==4.2
mock==2.0.0
pytest==3.0.2
pytest-cov==2.3.1
pytest-django==3.0.0
pytest-warnings==0.2.0
pytest-mock==1.2
django-nose==1.4.4
teamcity-messages==1.20
freezegun==0.3.7
moto==0.4.31